Fix a bug with scrollbars in Display Range command
It seems Java never reports the max value of a scrollbar all the
way to the end. You need to be considering getExtent() at the right
end and not considering it at the left end. Define some smoothish
interpolation values so that transition from x to x+10 near middle
of scrollbar is clean.